Redact sensitive information before sending
Sensitive data will now be redacted from collected k8s objects
on the collector-side rather than on the API-side.

The collector will also send a User-Agent header with every
request that contains the collector's version in the format
"k8s-collector vMAJOR.MINOR.PATCH".

To make it easier for the k8s-api to determine whether redaction
has occurred on data received, another new neader -
"X-Firefly-Redacted" is also added with the value "true".

The workflow and Dockerfile are updated to faciliate those new
requirements.

Relates to: INFL-7611